+++ Source/Core/CorridorGeometryLibrary.js
@@ -62,7 +62,7 @@ define([
     var quaterion = new Quaternion();
     var rotMatrix = new Matrix3();
     function computeRoundCorner (cornerPoint, startPoint, endPoint, cornerType, leftIsOutside, ellipsoid) {
-        var angle = Cartesian3.angleBetween(startPoint.subtract(cornerPoint, scratch1), endPoint.subtract(cornerPoint, scratch2));
+        var angle = Cartesian3.angleBetween(Cartesian3.subtract(startPoint, cornerPoint, scratch1), Cartesian3.subtract(endPoint, cornerPoint, scratch2));
         var granularity = (cornerType.value === CornerType.BEVELED.value) ? 1 : Math.ceil(angle / CesiumMath.toRadians(5)) + 1;
 
         var size = granularity * 3;
@@ -76,11 +76,11 @@ define([
         if (leftIsOutside) {
             m = Matrix3.fromQuaternion(Quaternion.fromAxisAngle(cornerPoint, angle / granularity, quaterion), rotMatrix);
         } else {
-            m = Matrix3.fromQuaternion(Quaternion.fromAxisAngle(cornerPoint.negate(scratch1), angle / granularity, quaterion), rotMatrix);
+            m = Matrix3.fromQuaternion(Quaternion.fromAxisAngle(Cartesian3.negate(cornerPoint, scratch1), angle / granularity, quaterion), rotMatrix);
         }
 
         var index = 0;
-        startPoint = startPoint.clone(scratch1);
+        startPoint = Cartesian3.clone(startPoint, scratch1);
         for ( var i = 0; i < granularity; i++) {
             startPoint = m.multiplyByVector(startPoint, startPoint);
             array[index++] = startPoint.x;
@@ -99,7 +99,7 @@ define([
         var leftEdge = calculatedPositions[1];
         startPoint = Cartesian3.fromArray(calculatedPositions[1], leftEdge.length - 3, startPoint);
         endPoint = Cartesian3.fromArray(calculatedPositions[0], 0, endPoint);
-        cornerPoint = startPoint.add(endPoint, cornerPoint).multiplyByScalar(0.5, cornerPoint);
+        cornerPoint = Cartesian3.multiplyByScalar(Cartesian3.add(startPoint, endPoint, cornerPoint), 0.5, cornerPoint);
         var firstEndCap = computeRoundCorner(cornerPoint, startPoint, endPoint, CornerType.ROUNDED, false, ellipsoid);
 
         var length = calculatedPositions.length - 1;
@@ -107,7 +107,7 @@ define([
         leftEdge = calculatedPositions[length];
         startPoint = Cartesian3.fromArray(rightEdge, rightEdge.length - 3, startPoint);
         endPoint = Cartesian3.fromArray(leftEdge, 0, endPoint);
-        cornerPoint = startPoint.add(endPoint, cornerPoint).multiplyByScalar(0.5, cornerPoint);
+        cornerPoint = Cartesian3.multiplyByScalar(Cartesian3.add(startPoint, endPoint, cornerPoint), 0.5, cornerPoint);
         var lastEndCap = computeRoundCorner(cornerPoint, startPoint, endPoint, CornerType.ROUNDED, false, ellipsoid);
 
         return [firstEndCap, lastEndCap];
@@ -118,7 +118,7 @@ define([
         if (leftIsOutside) {
             cornerPoint = Cartesian3.add(position, leftCornerDirection, cornerPoint);
         } else {
-            leftCornerDirection = leftCornerDirection.negate(leftCornerDirection);
+            leftCornerDirection = Cartesian3.negate(leftCornerDirection, leftCornerDirection);
             cornerPoint = Cartesian3.add(position, leftCornerDirection, cornerPoint);
         }
         return [cornerPoint.x, cornerPoint.y, cornerPoint.z, lastPoint.x, lastPoint.y, lastPoint.z];
@@ -127,19 +127,19 @@ define([
     function addShiftedPositions (positions, left, scalar, calculatedPositions) {
         var rightPositions = new Array(positions.length);
         var leftPositions = new Array(positions.length);
-        var scaledLeft = left.multiplyByScalar(scalar, scratch1);
-        var scaledRight = scaledLeft.negate(scratch2);
+        var scaledLeft = Cartesian3.multiplyByScalar(left, scalar, scratch1);
+        var scaledRight = Cartesian3.negate(scaledLeft, scratch2);
         var rightIndex = 0;
         var leftIndex = positions.length - 1;
 
         for (var i = 0; i < positions.length; i += 3) {
             var pos = Cartesian3.fromArray(positions, i, scratch3);
-            var rightPos = pos.add(scaledRight, scratch4);
+            var rightPos = Cartesian3.add(pos, scaledRight, scratch4);
             rightPositions[rightIndex++] = rightPos.x;
             rightPositions[rightIndex++] = rightPos.y;
             rightPositions[rightIndex++] = rightPos.z;
 
-            var leftPos = pos.add(scaledLeft, scratch4);
+            var leftPos = Cartesian3.add(pos, scaledLeft, scratch4);
             leftPositions[leftIndex--] = leftPos.z;
             leftPositions[leftIndex--] = leftPos.y;
             leftPositions[leftIndex--] = leftPos.x;
@@ -194,16 +194,16 @@ define([
         var position = positions[0]; //add first point
         var nextPosition = positions[1];
 
-        forward = Cartesian3.subtract(nextPosition, position, forward).normalize(forward);
+        forward = Cartesian3.normalize(Cartesian3.subtract(nextPosition, position, forward), forward);
         normal = ellipsoid.geodeticSurfaceNormal(position, normal);
-        left = normal.cross(forward, left).normalize(left);
+        left = Cartesian3.normalize(Cartesian3.cross(normal, forward, left), left);
         if (saveAttributes) {
             calculatedLefts.push(left.x, left.y, left.z);
             calculatedNormals.push(normal.x, normal.y, normal.z);
         }
         previousPos = Cartesian3.clone(position, previousPos);
         position = nextPosition;
-        backward = forward.negate(backward);
+        backward = Cartesian3.negate(forward, backward);
 
         var subdividedPositions;
         var corners = [];
@@ -212,19 +212,19 @@ define([
         for (i = 1; i < length - 1; i++) { // add middle points and corners
             normal = ellipsoid.geodeticSurfaceNormal(position, normal);
             nextPosition = positions[i + 1];
-            forward = Cartesian3.subtract(nextPosition, position, forward).normalize(forward);
-            cornerDirection = forward.add(backward, cornerDirection).normalize(cornerDirection);
-            var doCorner = !Cartesian3.equalsEpsilon(cornerDirection.negate(scratch1), normal, CesiumMath.EPSILON2);
+            forward = Cartesian3.normalize(Cartesian3.subtract(nextPosition, position, forward), forward);
+            cornerDirection = Cartesian3.normalize(Cartesian3.add(forward, backward, cornerDirection), cornerDirection);
+            var doCorner = !Cartesian3.equalsEpsilon(Cartesian3.negate(cornerDirection, scratch1), normal, CesiumMath.EPSILON2);
             if (doCorner) {
-                cornerDirection = cornerDirection.cross(normal, cornerDirection);
-                cornerDirection = normal.cross(cornerDirection, cornerDirection);
-                var scalar = width / Math.max(0.25, (Cartesian3.cross(cornerDirection, backward, scratch1).magnitude()));
+                cornerDirection = Cartesian3.cross(cornerDirection, normal, cornerDirection);
+                cornerDirection = Cartesian3.cross(normal, cornerDirection, cornerDirection);
+                var scalar = width / Math.max(0.25, Cartesian3.magnitude(Cartesian3.cross(cornerDirection, backward, scratch1)));
                 var leftIsOutside = angleIsGreaterThanPi(forward, backward, position, ellipsoid);
-                cornerDirection = cornerDirection.multiplyByScalar(scalar, cornerDirection, cornerDirection);
+                cornerDirection = Cartesian3.multiplyByScalar(cornerDirection, scalar, cornerDirection);
                 if (leftIsOutside) {
                     rightPos = Cartesian3.add(position, cornerDirection, rightPos);
-                    center = rightPos.add(left.multiplyByScalar(width, center), center);
-                    leftPos = rightPos.add(left.multiplyByScalar(width * 2, leftPos), leftPos);
+                    center = Cartesian3.add(rightPos, Cartesian3.multiplyByScalar(left, width, center), center);
+                    leftPos = Cartesian3.add(rightPos, Cartesian3.multiplyByScalar(left, width * 2, leftPos), leftPos);
                     scaleArray2[0] = Cartesian3.clone(previousPos, scaleArray2[0]);
                     scaleArray2[1] = Cartesian3.clone(center, scaleArray2[1]);
                     subdividedPositions = PolylinePipeline.scaleToSurface(scaleArray2, granularity, ellipsoid);
@@ -233,19 +233,19 @@ define([
                         calculatedLefts.push(left.x, left.y, left.z);
                         calculatedNormals.push(normal.x, normal.y, normal.z);
                     }
-                    startPoint = leftPos.clone(startPoint);
-                    left = normal.cross(forward, left).normalize(left);
-                    leftPos = rightPos.add(left.multiplyByScalar(width * 2, leftPos), leftPos);
-                    previousPos = rightPos.add(left.multiplyByScalar(width, previousPos), previousPos);
+                    startPoint = Cartesian3.clone(leftPos, startPoint);
+                    left = Cartesian3.normalize(Cartesian3.cross(normal, forward, left), left);
+                    leftPos = Cartesian3.add(rightPos, Cartesian3.multiplyByScalar(left, width * 2, leftPos), leftPos);
+                    previousPos = Cartesian3.add(rightPos, Cartesian3.multiplyByScalar(left, width, previousPos), previousPos);
                     if (cornerType.value === CornerType.ROUNDED.value || cornerType.value === CornerType.BEVELED.value) {
                         corners.push({leftPositions : computeRoundCorner(rightPos, startPoint, leftPos, cornerType, leftIsOutside, ellipsoid)});
                     } else {
-                        corners.push({leftPositions : computeMiteredCorner(position, startPoint, cornerDirection.negate(cornerDirection), leftPos, leftIsOutside, granularity, ellipsoid)});
+                        corners.push({leftPositions : computeMiteredCorner(position, startPoint, Cartesian3.negate(cornerDirection, cornerDirection), leftPos, leftIsOutside, granularity, ellipsoid)});
                     }
                 } else {
                     leftPos = Cartesian3.add(position, cornerDirection, leftPos);
-                    center = leftPos.add(left.multiplyByScalar(width, center).negate(center), center);
-                    rightPos = leftPos.add(left.multiplyByScalar(width * 2, rightPos).negate(rightPos), rightPos);
+                    center = Cartesian3.add(leftPos, Cartesian3.negate(Cartesian3.multiplyByScalar(left, width, center), center), center);
+                    rightPos = Cartesian3.add(leftPos, Cartesian3.negate(Cartesian3.multiplyByScalar(left, width * 2, rightPos), rightPos), rightPos);
                     scaleArray2[0] = Cartesian3.clone(previousPos, scaleArray2[0]);
                     scaleArray2[1] = Cartesian3.clone(center, scaleArray2[1]);
                     subdividedPositions = PolylinePipeline.scaleToSurface(scaleArray2, granularity, ellipsoid);
@@ -254,17 +254,17 @@ define([
                         calculatedLefts.push(left.x, left.y, left.z);
                         calculatedNormals.push(normal.x, normal.y, normal.z);
                     }
-                    startPoint = rightPos.clone(startPoint);
-                    left = normal.cross(forward, left).normalize(left);
-                    rightPos = leftPos.add(left.multiplyByScalar(width * 2, rightPos).negate(rightPos), rightPos);
-                    previousPos = leftPos.add(left.multiplyByScalar(width, previousPos).negate(previousPos), previousPos);
+                    startPoint = Cartesian3.clone(rightPos, startPoint);
+                    left = Cartesian3.normalize(Cartesian3.cross(normal, forward, left), left);
+                    rightPos = Cartesian3.add(leftPos, Cartesian3.negate(Cartesian3.multiplyByScalar(left, width * 2, rightPos), rightPos), rightPos);
+                    previousPos = Cartesian3.add(leftPos, Cartesian3.negate(Cartesian3.multiplyByScalar(left, width, previousPos), previousPos), previousPos);
                     if (cornerType.value === CornerType.ROUNDED.value || cornerType.value === CornerType.BEVELED.value) {
                         corners.push({rightPositions : computeRoundCorner(leftPos, startPoint, rightPos, cornerType, leftIsOutside, ellipsoid)});
                     } else {
                         corners.push({rightPositions : computeMiteredCorner(position, startPoint, cornerDirection, rightPos, leftIsOutside, granularity, ellipsoid)});
                     }
                 }
-                backward = forward.negate(backward);
+                backward = Cartesian3.negate(forward, backward);
             }
             position = nextPosition;
         }