+++ Specs/Scene/CameraFlightPathSpec.js
@@ -197,13 +197,13 @@ defineSuite([
     it('creates an animation in 3d', function() {
         var camera = frameState.camera;
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.negate();
-        var endDirection = startDirection.negate();
-        var endUp = startUp.negate();
+        var endPosition = Cartesian3.negate(startPosition);
+        var endDirection = Cartesian3.negate(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimation(scene, {
@@ -227,14 +227,14 @@ defineSuite([
     it('creates an animation in 3d with cartographic', function() {
         var camera = frameState.camera;
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.negate();
+        var endPosition = Cartesian3.negate(startPosition);
         var endCartographic = frameState.scene2D.projection.getEllipsoid().cartesianToCartographic(endPosition);
-        var endDirection = startDirection.negate();
-        var endUp = startUp.negate();
+        var endDirection = Cartesian3.negate(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimationCartographic(scene, {
@@ -260,14 +260,14 @@ defineSuite([
 
         var startPosition = frameState.scene2D.projection.getEllipsoid().cartographicToCartesian(new Cartographic(CesiumMath.PI, 0, 20));
         camera.position = startPosition;
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.negate();
+        var endPosition = Cartesian3.negate(startPosition);
         var endCartographic = frameState.scene2D.projection.getEllipsoid().cartesianToCartographic(endPosition);
         var extent = new Extent(endCartographic.longitude - 0.0000019, endCartographic.latitude - 0.0000019, endCartographic.longitude + 0.0000019, endCartographic.latitude + 0.0000019);
-        var endDirection = startDirection.negate();
-        var endUp = startUp.negate();
+        var endDirection = Cartesian3.negate(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimationExtent(scene, {
@@ -293,17 +293,17 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimation(scene, {
@@ -329,18 +329,18 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
         var endCartographic = frameState.scene2D.projection.unproject(endPosition);
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimationCartographic(scene, {
@@ -368,18 +368,18 @@ defineSuite([
 
         var startPosition = frameState.scene2D.projection.getEllipsoid().cartographicToCartesian(new Cartographic(CesiumMath.PI, 0, 20));
         camera.position = startPosition;
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
 
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 20.0));
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 20.0));
         var endCartographic = frameState.scene2D.projection.unproject(endPosition);
         var extent = new Extent(endCartographic.longitude - 0.0000019, endCartographic.latitude - 0.0000019, endCartographic.longitude + 0.0000019, endCartographic.latitude + 0.0000019);
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimationExtent(scene, {
@@ -405,19 +405,19 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
         camera.frustum = createOrthographicFrustum();
 
         var startHeight = camera.frustum.right - camera.frustum.left;
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimation(scene, {
@@ -447,20 +447,20 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
         camera.frustum = createOrthographicFrustum();
 
         var startHeight = camera.frustum.right - camera.frustum.left;
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
         var endCartographic = frameState.scene2D.projection.unproject(endPosition);
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimationCartographic(scene, {
@@ -491,21 +491,21 @@ defineSuite([
         camera.controller._mode = SceneMode.SCENE2D;
 
         camera.position = new Cartesian3(CesiumMath.PI, 0.0, 20.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
         camera.frustum = createOrthographicFrustum();
 
         var startHeight = camera.frustum.right - camera.frustum.left;
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 0.0));
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 0.0));
         var endCartographic = frameState.scene2D.projection.unproject(endPosition);
         var extent = new Extent(endCartographic.longitude - 0.0000019, endCartographic.latitude - 0.0000019, endCartographic.longitude + 0.0000019, endCartographic.latitude + 0.0000019);
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var duration = 5000.0;
         var flight = CameraFlightPath.createAnimationExtent(scene, {
@@ -531,7 +531,7 @@ defineSuite([
 
     it('creates a path where the start and end points only differ in height', function() {
         var camera = frameState.camera;
-        var start = camera.position.clone();
+        var start = Cartesian3.clone(camera.position);
         var end = Ellipsoid.WGS84.cartesianToCartographic(start);
         end.height -= 1000000.0;
 
@@ -552,13 +552,13 @@ defineSuite([
     it('does not create a path to the same point', function() {
         var camera = frameState.camera;
         camera.position = new Cartesian3(7000000.0, 0.0, 0.0);
-        camera.direction = camera.position.normalize().negate();
-        camera.up = Cartesian3.UNIT_Z.clone();
-        camera.right = camera.direction.cross(camera.up).normalize();
+        camera.direction = Cartesian3.negate(Cartesian3.normalize(camera.position));
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Z);
+        camera.right = Cartesian3.normalize(Cartesian3.cross(camera.direction, camera.up));
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
         var duration = 3000.0;
         var flight = CameraFlightPath.createAnimation(scene, {
@@ -600,12 +600,12 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
         camera.frustum = createOrthographicFrustum();
         var frustum = camera.frustum;
-        var destination = camera.position.clone();
+        var destination = Cartesian3.clone(camera.position);
         destination.z = Math.max(frustum.right - frustum.left, frustum.top - frustum.bottom);
 
         var flight = CameraFlightPath.createAnimation(scene, {
@@ -620,9 +620,9 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
         camera.frustum = createOrthographicFrustum();
 
         var flight = CameraFlightPath.createAnimation(scene, {
@@ -637,9 +637,9 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
 
         var flight = CameraFlightPath.createAnimation(scene, {
             destination : camera.position
@@ -653,18 +653,18 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
         camera.frustum = createOrthographicFrustum();
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
-        var endDirection = startDirection.clone();
-        var endUp = startUp.negate();
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
+        var endDirection = Cartesian3.clone(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var flight = CameraFlightPath.createAnimation(scene, {
             destination : endPosition,
@@ -687,12 +687,12 @@ defineSuite([
         var camera = frameState.camera;
 
         camera.position = new Cartesian3(0.0, 0.0, 1000.0);
-        camera.direction = Cartesian3.UNIT_Z.negate();
-        camera.up = Cartesian3.UNIT_Y.clone();
-        camera.right = camera.direction.cross(camera.up);
+        camera.direction = Cartesian3.negate(Cartesian3.UNIT_Z);
+        camera.up = Cartesian3.clone(Cartesian3.UNIT_Y);
+        camera.right = Cartesian3.cross(camera.direction, camera.up);
 
-        var startPosition = camera.position.clone();
-        var endPosition = startPosition.add(new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
+        var startPosition = Cartesian3.clone(camera.position);
+        var endPosition = Cartesian3.add(startPosition, new Cartesian3(-6e6 * Math.PI, 6e6 * CesiumMath.PI_OVER_FOUR, 100.0));
 
         var flight = CameraFlightPath.createAnimation(scene, {
             destination : endPosition,
@@ -707,13 +707,13 @@ defineSuite([
     it('creates an animation in 3d 0 duration', function() {
         var camera = frameState.camera;
 
-        var startPosition = camera.position.clone();
-        var startDirection = camera.direction.clone();
-        var startUp = camera.up.clone();
+        var startPosition = Cartesian3.clone(camera.position);
+        var startDirection = Cartesian3.clone(camera.direction);
+        var startUp = Cartesian3.clone(camera.up);
 
-        var endPosition = startPosition.negate();
-        var endDirection = startDirection.negate();
-        var endUp = startUp.negate();
+        var endPosition = Cartesian3.negate(startPosition);
+        var endDirection = Cartesian3.negate(startDirection);
+        var endUp = Cartesian3.negate(startUp);
 
         var flight = CameraFlightPath.createAnimation(scene, {
             destination : endPosition,